Removed pupil register v1 implementation and config and updated hdf pupil status related calls to pupil status service
