+++ admin/services/headteacher-declaration.service.js
@@ -1,42 +1,17 @@
 'use strict'
 
 const moment = require('moment-timezone')
-const R = require('ramda')
 
 const config = require('../config')
 const schoolDataService = require('../services/data-access/school.data.service')
 const checkWindowV2Service = require('../services/check-window-v2.service')
 const attendanceCodeDataService = require('./data-access/attendance-code.data.service')
 const pupilAttendanceDataService = require('../services/data-access/pupil-attendance.data.service')
+const pupilStatusService = require('../services/pupil-status.service')
 const headteacherDeclarationDataService = require('./data-access/headteacher-declaration.data.service')
+const pupilStatusDataService = require('./data-access/pupil-status.data.service')
 const headteacherDeclarationService = {}
-const pupilRegisterService = require('./pupil-register.service')
 const settingService = require('./setting.service')
-
-/**
- * Add the pupil `status` field, using the pupil register service to do so.
- * @param {Object} settings
- * @param {Object} pupil
- * @return {Object} partially cloned pupil obj with an additional `status` property
- */
-headteacherDeclarationService.addStatus = function addStatus (settings, pupil) {
-  const newPupil = R.pickAll(['id', 'foreName', 'lastName', 'middleNames', 'dateOfBirth', 'urlSlug', 'checkStatusCode',
-    'group_id', 'reason', 'reasonCode'], pupil)
-  newPupil.status = pupilRegisterService.getProcessStatusV2({
-    attendanceId: pupil.attendanceId,
-    checkComplete: pupil.checkComplete,
-    checkReceived: pupil.checkReceived,
-    checkStatusCode: pupil.checkStatusCode,
-    currentCheckId: pupil.currentCheckId,
-    notReceivedExpiryInMinutes: settings.checkTimeLimit,
-    pinExpiresAt: pupil.pinExpiresAt,
-    pupilCheckComplete: pupil.pupilCheckComplete,
-    pupilId: pupil.pupilId,
-    pupilLoginDate: pupil.pupilLoginDate,
-    restartAvailable: pupil.restartAvailable
-  })
-  return newPupil
-}
 
 /**
  * @typedef {Object} hdfPupil
@@ -61,21 +36,14 @@ headteacherDeclarationService.findPupilsForSchool = async (schoolId) => {
   if (!schoolId) {
     throw new Error('schoolId is required')
   }
-
-  const settings = await settingService.get()
-
-  // We can get pupils and attendance codes directly from the database...
-  const pupils = await headteacherDeclarationDataService.sqlFindPupilsFullStatus(schoolId)
-  // ...and add the HDF status in.  The HDF status is the same as the pupil outcome status on the pupil register / status
-  // screen, but the Non-attendance reason shown in the status, rather than 'Not attending'.
-  const pupilsWithStatus = pupils.map(R.partial(headteacherDeclarationService.addStatus, [settings]))
-  return pupilsWithStatus
+  // The HDF status is the same as the pupil outcome status on the pupil register / status screen
+  return pupilStatusService.getPupilStatusData(schoolId)
 }
 
 /**
  * Find the a pupil for the given pupilId and dfeNumber
- * @param pupilId
- * @param dfeNumber
+ * @param urlSlug
+ * @param schoolId
  * @return {Promise<object>}
  */
 headteacherDeclarationService.findPupilBySlugAndSchoolId = async function findPupilBySlugAndSchoolId (urlSlug, schoolId) {
@@ -86,9 +54,8 @@ headteacherDeclarationService.findPupilBySlugAndSchoolId = async function findPu
     throw new Error('schoolId param is required')
   }
   const settings = await settingService.get()
-  const pupil = await headteacherDeclarationDataService.sqlFindOnePupilFullStatus(urlSlug, schoolId)
-  const newPupil = this.addStatus(settings, pupil)
-  return newPupil
+  const pupil = await pupilStatusDataService.sqlFindOnePupilFullStatus(urlSlug, schoolId)
+  return pupilStatusService.addStatus(settings, pupil)
 }
 
 /**
